some fixes

+++ crates/cubecl/benches/matmul.rs
@@ -207,6 +207,6 @@ fn main() {
     }
     #[cfg(feature = "wgpu-msl")]
     {
-        run_benches::<cubecl::wgpu::WgpuRuntime, half::f16>();
+        run_benches::<cubecl::wgpu::WgpuRuntime, f32>();
     }
 }
